Fix: error when Basic auth is required and receive Bearer token
