+++ lib/middlewares/accessTokenParser.js
@@ -1,32 +1,10 @@
 const debug = require('debug')('crowi:middlewares:accessTokenParser')
 
-/**
- * Extract Bearer token from Authorization header.
- *
- * @param {Object} headers HTTP request headers
- * @param {string} headers.authorization Authorization header value.
- * @return {?string} found access_token or null.
- */
-const extractBearerToken = headers => {
-  const v = headers.authorization
-  if (!v) {
-    return null
-  }
-
-  const parts = v
-    .trim()
-    .replace(/( )+/g, ' ')
-    .split(' ')
-  if (parts.length !== 2 || parts[0] !== 'Bearer') {
-    return null
-  }
-
-  return parts[1]
-}
+const { parseAccessToken } = require('../util/accessTokenParser')
 
 module.exports = (crowi, app) => {
   return (req, res, next) => {
-    var accessToken = extractBearerToken(req.headers) || req.query.access_token || req.body.access_token || null
+    const accessToken = parseAccessToken(req)
     if (!accessToken) {
       return next()
     }
