+++ lib/middlewares/basicAuth.js
@@ -1,9 +1,12 @@
 const basicAuth = require('basic-auth-connect')
 
+const { parseAccessToken } = require('../util/accessTokenParser')
+
 module.exports = (crowi, app) => {
   return (req, res, next) => {
-    var config = crowi.getConfig()
-    if (req.query.access_token || req.body.access_token) {
+    const config = crowi.getConfig()
+    const accessToken = parseAccessToken(req)
+    if (accessToken) {
       return next()
     }
